Prevent inlining of hpat_get_set_since_epoch.
Add checkpoint start, write and end functions so that their return types are correct but these are pattern matched and
replaced by hpat later.

Add the calls to start and end checkpoint as well as the code to checkpoint individual variables.
Do a hack to multiply define the param to start checkpoint so that the function won't be hoisted above the conditional.